+++ R/nback.R
@@ -41,25 +41,11 @@ nback <- function(data, ...) {
       pc = mean(.data$acc_adj == 1),
       mrt = mean(.data$RT[.data$acc_adj == 1])
     )
-  sdt <- data_adj %>%
-    dplyr::group_by(.data$type_adj) %>%
-    dplyr::summarise(
-      n = dplyr::n(),
-      pc = mean(.data$acc_adj == 1)
-    ) %>%
-    dplyr::mutate(
-      pc_adj = dplyr::case_when(
-        .data$pc == 0 ~ 1 / (2 * .data$n),
-        .data$pc == 1 ~ 1 - 1 / (2 * .data$n),
-        TRUE ~ .data$pc
-      )
-    ) %>%
-    dplyr::select(.data$type_adj, .data$pc_adj) %>%
-    tidyr::pivot_wider(names_from = "type_adj", values_from = "pc_adj") %>%
-    dplyr::transmute(
-      dprime = stats::qnorm(.data$s) + stats::qnorm(.data$n),
-      c = -(stats::qnorm(.data$s) - stats::qnorm(.data$n)) / 2
-    )
+  sdt <- calc_sdt(
+    data_adj,
+    name_type = "type_adj",
+    name_acc = "acc_adj"
+  )
   is_normal <- data_adj %>%
     dplyr::summarise(nt = dplyr::n(), nc = sum(.data$acc_adj == 1)) %>%
     dplyr::transmute(is_normal = .data$nc > stats::qbinom(0.95, .data$nt, 0.5))