+++ R/utils-calculation.R
@@ -106,55 +106,41 @@ calc_switch_cost <- function(data,
     )
   cbind(switch_cost_count, switch_cost_rt)
 }
-#' Get the matched variables for input data
+#' Signal Detection Theory
 #'
-#' Input data require several variables to be existed, and some variables can
-#' have alternatives. Here we check if all the variables exist, and if only one
-#' of the alternatives exists for each variable. If the check is passed, the
-#' matched result (a character vector) will be returned, or else a `NULL` value
-#' will be returned.
+#' Utility function to calculate indices related to signal detection theory,
+#' e.g., d', c.
 #'
-#' @param data Required. Raw data, a `data.frame`.
-#' @param vars_config Required. A `data.frame` containing the configuration of
-#'   required variables. Possible variables are:
-#'   * `field`: Required. Name references of all data variable names.
-#'   * `name`: Required. Data variable names, in which lists all the possible
-#'   names for current data variable.
-#'   * `type`: Optional. Can be "required" (must exist) or "optional" (try to
-#'     match, if not find, `NA` is returned and it must be handled in another
-#'     place). If this variable is not specified, it will default to "required".
+#' @param data Required. A `data.frame` from which the indices are calculated.
+#' @param name_type The name of the variable in `data` storing the signal type
+#'   info.
+#' @param name_acc The name of the variable in `data` storing accuracy data.
+#' @param values_type The values of signal type. The first is 'signal', and the
+#'   second is 'noise'.
 #' @keywords internal
-match_data_vars <- function(data, vars_config) {
-  # set default type to "required"
-  if (!utils::hasName(vars_config, "type")) {
-    vars_config$type <- "required"
-  }
-  vars_chk_result <- vars_config %>%
+calc_sdt <- function(data,
+                     name_type = "Type",
+                     name_acc = "ACC",
+                     values_type = c("s", "n")) {
+  sdt <- data %>%
+    dplyr::group_by(.data[[name_type]]) %>%
+    dplyr::summarise(
+      n = dplyr::n(),
+      pc = mean(.data[[name_acc]] == 1)
+    ) %>%
     dplyr::mutate(
-      purrr::map_df(
-        .data$name,
-        ~ {
-          match_result <- utils::hasName(data, .x)
-          tibble(
-            is_found = any(match_result),
-            is_confused = sum(match_result) > 1,
-            # the match can be of length 0, so use `if` instead of `if_else`
-            matched = if (sum(match_result) == 1) .x[match_result] else NA
-          )
-        }
+      pc_adj = dplyr::case_when(
+        .data$pc == 0 ~ 1 / (2 * .data$n),
+        .data$pc == 1 ~ 1 - 1 / (2 * .data$n),
+        TRUE ~ .data$pc
       )
+    ) %>%
+    dplyr::select(.data[[name_type]], .data$pc_adj) %>%
+    tidyr::pivot_wider(names_from = "type_adj", values_from = "pc_adj") %>%
+    dplyr::transmute(
+      dprime = stats::qnorm(.data[[values_type[[1]]]]) +
+        stats::qnorm(.data[[values_type[[2]]]]),
+      c = -(stats::qnorm(.data[[values_type[[1]]]]) -
+        stats::qnorm(.data[[values_type[[2]]]])) / 2
     )
-  if (!all(vars_chk_result$is_found[vars_chk_result$type == "required"])) {
-    warning("At least one of the required variables are missing.")
-    return(NULL)
-  }
-  if (any(vars_chk_result$is_confused)) {
-    warning("At least one of the variables have more than one match.")
-    return(NULL)
-  }
-  return(
-    vars_chk_result %>%
-      dplyr::select(.data$field, .data$matched) %>%
-      tibble::deframe()
-  )
 }