+++ NEWS.md
@@ -1,3 +1,8 @@
+# dataprocr2 0.1.4
+
+* Added a utility function `calc_sdt()` to calculate signal detection theory related indices.
+* Fixed an issue related to perfect accuracy in `drm()`.
+
 # dataprocr2 0.1.3
 
 * Added a `NEWS.md` file to track changes to the package.